+++ lib/packages/core/src/api/factories/asset/index.ts
@@ -1,3 +1,4 @@
 /** @module core.api.factories */
 
 export * from './getAsset';
+export * from './getAllAssets';